refactor: split css-selector-parser into modules
